Put offer_pooling_intervall in config
