+++ golem/clientconfigdescriptor.py
@@ -39,6 +39,7 @@ class ClientConfigDescriptor(object):
         self.resource_session_timeout = 0
         self.clean_resources_older_than_seconds = 0
         self.clean_tasks_older_than_seconds = 0
+        self.offer_pooling_interval = 0.0
 
         self.node_snapshot_interval = 0.0
         self.network_check_interval = 0.0
