+++ golem/marketplace/rust.py
@@ -1,0 +1,1 @@
+from rust.golem import marketplace__order_providers as order_providers  # noqa pylint: disable=no-name-in-module,import-error
